% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_campaigns.R
\name{pardot_campaigns}
\alias{pardot_campaigns}
\title{Retrieve Pardot Campaigns}
\usage{
pardot_campaigns(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, id_greater_than, id_less_than, name, 
updated_before, updated_after.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#campaign.
}
\description{
Make a call to the Pardot API and return the campaigns matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_campaigns()
df <- pardot_campaigns(created_after = 'today')}
}
