% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_client.R
\name{pardot_client}
\alias{pardot_client}
\title{Make a call to the Pardot API and return XML or data frame}
\usage{
pardot_client(object, operator, identifier_field = NULL, identifier = NULL,
  request_pars = NULL, result_format = "json", unlist_dataframe = TRUE,
  verbose = 0)
}
\arguments{
\item{object}{A string containing a Pardot Object}

\item{operator}{A string containing a Pardot Operator}

\item{identifier_field}{A string with an optional identifier field. Can be null}

\item{identifier}{A string with an optional identifier that can be null if identifier_field is null}

\item{request_pars}{A string of query parameters. Can be null}

\item{result_format}{A string specifying the result format used for API calls: "json" (default) or "xml". If json, pardot_client() returns a data frame.}

\item{unlist_dataframe}{A logical, default TRUE. If it is FALSE all fields having embedded lists are returned as they are. If unlist_dataframe is TRUE a field with embedded list(s) is converted to multiple records and/or fields. The values of the other fields are duplicated across these records. Applies to object "visit".}

\item{verbose}{Verbose output. Integer value, default zero is non-verbose. 1 displays a progress bar consisting of dots and numbers, for respectively every 200th and 1000th received record. 2 displays the successive call urls and the data structure returned by the first call.}
}
\value{
XML or a data frame.
}
\description{
Make a call to the Pardot API and return XML or data frame
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
pardot_client("campaign", "query")
pardot_client(object = "campaign", operator = "query", 
  request_params = "created_after=yesterday&id_greater_than=492276479")}
}
