% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_email.R
\name{pardot_email}
\alias{pardot_email}
\title{Retrieve Pardot Email}
\usage{
pardot_email(list_email_id, verbose = 0, ...)
}
\arguments{
\item{list_email_id}{The Pardot ID of the target email.}

\item{verbose}{Verbose output. See pardot_client().}

\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are include_message.}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#email.
}
\description{
Make a call to the Pardot API and return the data for the specified email.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_emaildata(email_id = 747447245)}
}
