% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_email_clicks.R
\name{pardot_email_clicks}
\alias{pardot_email_clicks}
\title{Retrieve Pardot Email Clicks}
\usage{
pardot_email_clicks(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, id_greater_than, list_email_id, drip_program_action_id, email_template_id, tracker_redirect_id.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#email-clicks.
}
\description{
Make a call to the Pardot API and return the email clicks matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_email_clicks()
df <- pardot_email_clicks(created_after = 'yesterday') }
}
