% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_email_stats.R
\name{pardot_email_stats}
\alias{pardot_email_stats}
\title{Retrieve Pardot Email Stats}
\usage{
pardot_email_stats(list_email_id, verbose = 0, ...)
}
\arguments{
\item{list_email_id}{The Pardot ID of the target email.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame.
}
\description{
Make a call to the Pardot API and return the statistical data for the specified email.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_email_stats(list_email_id = 747447245)}
}
