% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_email_template.R
\name{pardot_email_template}
\alias{pardot_email_template}
\title{Retrieve Pardot Email Template}
\usage{
pardot_email_template(email_template_id, verbose = 0)
}
\arguments{
\item{email_template_id}{The Pardot ID of the target email template.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame.
}
\description{
Make a call to the Pardot API and return the data for the specified email template.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_email_template(email_template_id = 21918)}
}
