% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_lifecycle_histories.R
\name{pardot_lifecycle_histories}
\alias{pardot_lifecycle_histories}
\title{Retrieve Pardot Lifecycle Histories}
\usage{
pardot_lifecycle_histories(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, id_greater_than, id_less_than.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#lifecycle-history.
}
\description{
Make a call to the Pardot API and return the lifecycle histories matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_lifecycle_histories()
df <- pardot_lifecycle_histories(id_greater_than = '123')}
}
