% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_lifecycle_stages.R
\name{pardot_lifecycle_stages}
\alias{pardot_lifecycle_stages}
\title{Retrieve Pardot Lifecycle Stages}
\usage{
pardot_lifecycle_stages(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are id_greater_than, id_less_than.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#lifecycle-stage.
}
\description{
Make a call to the Pardot API and return the lifecycle stages matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_lifecycle_stages()
df <- pardot_lifecycle_stages(id_greater_than = '123')}
}
