% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_list_memberships.R
\name{pardot_list_memberships}
\alias{pardot_list_memberships}
\title{Retrieve Pardot List Memberships}
\usage{
pardot_list_memberships(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, deleted, id_greater_than, 
id_less_than, list_id, updated_before, updated_after.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#list-membership.
}
\description{
Make a call to the Pardot API and return the list memberships matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_list_memberships()
df <- pardot_list_memberships(created_after = 'today')}
}
