% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_prospect_accounts.R
\name{pardot_prospect_accounts}
\alias{pardot_prospect_accounts}
\title{Retrieve Pardot Prospect Accounts}
\usage{
pardot_prospect_accounts(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, id_greater_than, id_less_than, name, updated_after, updated_before.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#prospect-account.
}
\description{
Make a call to the Pardot API and return the prospects matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_prospect_accounts()
df <- pardot_prospect_accounts(created_after = 'today')}
}
