% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_prospects.R
\name{pardot_prospects}
\alias{pardot_prospects}
\title{Retrieve Pardot Prospects}
\usage{
pardot_prospects(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are assigned, assigned_to_user, created_after, created_before, deleted, grade_equal_to,
grade_greater_than, grade_less_than, id_greater_than, id_less_than, is_starred, last_activity_before, 
last_activity_after, last_activity_never, list_id, new, score_equal_to, score_greater_than, score_less_than, 
updated_after, updated_before.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#prospect.
}
\description{
Make a call to the Pardot API and return the prospects matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_prospects()
df <- pardot_prospects(created_after = 'today')}
}
