% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_tag_objects.R
\name{pardot_tag_objects}
\alias{pardot_tag_objects}
\title{Retrieve Pardot Tag Objects}
\usage{
pardot_tag_objects(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, id_greater_than, id_less_than.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#tag-object.
}
\description{
Make a call to the Pardot API and return the tag objects matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_tag_objects()
df <- pardot_tag_objects(created_after = 'yesterday') }
}
