% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_visitor_activities.R
\name{pardot_visitor_activities}
\alias{pardot_visitor_activities}
\title{Retrieve Pardot Visitor Activities}
\usage{
pardot_visitor_activities(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are created_after, created_before, id_greater_than, id_less_than, prospect_only, type,
custom_url_only, email_only, file_only, form_only, form_handler_only, landing_page_only, campaign_id, 
custom_url_id, email_id, file_id, form_id, form_handler_id, landing_page_id, prospect_id, visitor_id.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#visitor-activity.
}
\description{
Make a call to the Pardot API and return the visitor activities matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_visitor_activity()
df <- pardot_visitor_activity(created_after = 'yesterday', prospect_only = 'true') }
}
