% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pardot_visits.R
\name{pardot_visits}
\alias{pardot_visits}
\title{Retrieve Pardot Visits}
\usage{
pardot_visits(..., verbose = 0)
}
\arguments{
\item{...}{Comma separated list of parameter name and parameter value pairs. Parameter names are not quoted. 
Allowed parameter names are ids, visitor_ids, prospect_ids.}

\item{verbose}{Verbose output. See pardot_client().}
}
\value{
A data frame. See http://developer.pardot.com/kb/object-field-references/#visit and http://developer.pardot.com/kb/object-field-references/#visitor-page-view. Note: The field visitor_page_view in the returned data frame contains nested lists. Use pardot_client() parameter unlist_dataframe = TRUE to expand the list field into data frame rows, or unlist_dataframe = FALSE to retain the lists as they are.
}
\description{
Make a call to the Pardot API and returns the visits matching the specified criteria parameters.
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")
df <- pardot_visits()
df <- pardot_visits(created_after = 'yesterday', prospect_id = 123) }
}
