% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_credentials.R
\name{set_credentials}
\alias{set_credentials}
\title{Set the Pardot user credentials}
\usage{
set_credentials(pardot_username, pardot_password, pardot_user_key)
}
\arguments{
\item{pardot_username}{A string containing your Pardot UserName}

\item{pardot_password}{A string containing your Pardot Password}

\item{pardot_user_key}{A string containing your Pardot User Key}
}
\description{
Simple function to set the users pardot credentials in an R environment variable
}
\examples{
\dontrun{
set_credentials("your-username", "your-password", "your-user-key")}
}
